moon_manager/requirements.txt
hug!=2.5.0
pyyaml
moon_utilities
moon_engine
tinydb
gunicorn
requests
sqlalchemy
pymysql
